feat: added stories for Tab Component
