--- src/components/base/Tab.mocks.ts
@@ -1,11 +1,0 @@
-import { ITab } from './Tab';
-
-const base: ITab = {
-  text: 'short break',
-  tabID: 'short-break',
-  isSelected: true,
-};
-
-export const mockTabProps = {
-  base,
-};
