+++ src/components/tab/Tab.module.css
@@ -21,8 +21,3 @@ label {
 label:hover {
   box-shadow: inset 0 0 100px 100px rgba(255, 255, 255, 0.25);
 }
-
-/*
-hidden peer/pomodoro
-
-cursor-pointer transition-text transition-colors peer-checked/pomodoro:text-[#1E213F] peer-checked/pomodoro:bg-[#F87070] tab z-2 text-[#D7E0FF] transition-text rounded-full px-4 py-2 */
